[BIB-39] Make deadline value an amount of pages, not a percent

+++ apps/readings/views.py
@@ -29,9 +29,10 @@ def set_deadline(request, reading_id: int):
         return HttpResponse(status=http_status.HTTP_400_BAD_REQUEST)
 
     try:
-        deadline_percent = int(request.POST.get("percent", 0))
+        deadline_percent = float(request.POST.get("percent", 0))
     except ValueError:
         deadline_percent = 100
+
 
     if not (0 <= deadline_percent <= 100):
         deadline_percent = 100
